formatted auto classes in diff folders, LEDs
