+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/encodedautos/pushBlueB.java
@@ -1,6 +1,5 @@
-package org.firstinspires.ftc.teamcode.autos;
+package org.firstinspires.ftc.teamcode.encodedautos;
 
-import com.acmerobotics.roadrunner.drive.Drive;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 
@@ -10,8 +9,8 @@ import org.firstinspires.ftc.teamcode.main.Encoded;
 import org.firstinspires.ftc.teamcode.vision.FirstVisionProcessor;
 import org.firstinspires.ftc.vision.VisionPortal;
 
-@Autonomous(name="red Front")
-public class redFront extends OpMode {
+@Autonomous(name="PUSH blue Back")
+public class pushBlueB extends OpMode {
     private FirstVisionProcessor visionProcessor;
     private VisionPortal visionPortal;
     private Encoded encoded;
@@ -36,25 +35,30 @@ public class redFront extends OpMode {
         telemetry.addData("Identified", visionProcessor.getSelection());
         switch (visionProcessor.getSelection()) {
             case LEFT:
-                driveTrain.forward(28,700);
-                driveTrain.turnLeft(23,700);
-                driveTrain.forward(4,700);
-                driveTrain.backward(2,700);
+                encoded.closeClaw();
+                driveTrain.backward(23,800);
+                driveTrain.strafeRight(12,800);
+                driveTrain.backward(4,800);
+                driveTrain.forward(8,500);
                 break;
 
             case NONE:
             case MIDDLE:
-                driveTrain.forward(30,900);
-                driveTrain.backward(3,900);
+                encoded.closeClaw();
+                driveTrain.backward(30,800);
+                driveTrain.forward(4,800);
+                driveTrain.turnRight(19.5,700);
+                driveTrain.forward(10,800);
+                driveTrain.strafeRight(2,800);
                 break;
 
             case RIGHT:
-                driveTrain.forward(28,800);
-                driveTrain.turnRight(25,900);
-                driveTrain.forward(5,700);
-                driveTrain.backward(2,700);
+                encoded.closeClaw();
+                driveTrain.backward(29,800);
+                driveTrain.turnRight(20,300);
+                driveTrain.backward(3,700);
+                driveTrain.forward(4,800);
                 break;
-
         }
     }
