+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/rrautos/rrBlueBackGATE.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.autos;
+package org.firstinspires.ftc.teamcode.rrautos;
 
 import com.acmerobotics.roadrunner.geometry.Pose2d;
 import com.acmerobotics.roadrunner.geometry.Vector2d;
@@ -12,7 +12,7 @@ import org.firstinspires.ftc.teamcode.main.Encoded;
 import org.firstinspires.ftc.teamcode.vision.FirstVisionProcessor;
 import org.firstinspires.ftc.vision.VisionPortal;
 
-@Autonomous(name = "rrBlueBackGATE")
+@Autonomous(name = "RR BlueBackGATE")
 public class rrBlueBackGATE extends OpMode {
     private FirstVisionProcessor visionProcessor;
     private VisionPortal visionPortal;
