+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/rrautos/rrBlueBackYellowOnly.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.autos;
+package org.firstinspires.ftc.teamcode.rrautos;
 
 import com.acmerobotics.roadrunner.geometry.Pose2d;
 import com.acmerobotics.roadrunner.geometry.Vector2d;
@@ -11,8 +11,8 @@ import org.firstinspires.ftc.teamcode.main.Encoded;
 import org.firstinspires.ftc.teamcode.vision.FirstVisionProcessor;
 import org.firstinspires.ftc.vision.VisionPortal;
 
-@Autonomous(name = "rrBlueFrontTRUSSyellowOnly")
-public class rrBlueFrontTRUSSyellowOnly extends OpMode {
+@Autonomous(name = "RR BlueBackYellowOnly")
+public class rrBlueBackYellowOnly extends OpMode {
     private FirstVisionProcessor visionProcessor;
     private VisionPortal visionPortal;
     private Encoded encoded;
@@ -23,7 +23,6 @@ public class rrBlueFrontTRUSSyellowOnly extends OpMode {
         visionProcessor = new FirstVisionProcessor();
         visionPortal = VisionPortal.easyCreateWithDefaults(hardwareMap.get(WebcamName.class, "Webcam 1"), visionProcessor);
     }
-
     @Override
     public void init_loop() {
         telemetry.addData("Identified", visionProcessor.getSelection());
@@ -36,23 +35,18 @@ public class rrBlueFrontTRUSSyellowOnly extends OpMode {
         telemetry.addData("Identified", visionProcessor.getSelection());
         switch (visionProcessor.getSelection()) {
             case LEFT:
-                drive.trajectorySequenceBuilder(new Pose2d(-35, 70, Math.toRadians(90)))
-                        .lineToConstantHeading(new Vector2d(-40, 50)) // positioning
-                        .lineToLinearHeading(new Pose2d(-37, 29, Math.toRadians(180))) // orientation
-                        .lineToConstantHeading(new Vector2d(-35, 29)) //slow push to spikemark
+                drive.trajectorySequenceBuilder(new Pose2d(12, 74, Math.toRadians(90)))
+                        .lineToConstantHeading(new Vector2d(23,35)) // to spikemark
                         .addTemporalMarker(0,()-> {encoded.armtoGround();})
                         .addTemporalMarker(0.5,()->{encoded.openBottomClaw();})
-                        .lineToConstantHeading(new Vector2d(-38, 29)) // safe backup
-                        .lineToConstantHeading(new Vector2d(-42, 58.5)) // orientate + line up for truss
-                        .lineToConstantHeading(new Vector2d(42, 58.5)) // fly under truss
-                        .splineToLinearHeading(new Pose2d(50, 35.5), Math.toRadians(0)) // to bd
+                        .lineToConstantHeading(new Vector2d(23, 40)) // back up
+                        .splineToLinearHeading(new Pose2d(50,42), Math.toRadians(0)) // to bd
                         .addTemporalMarker(0,()-> {encoded.armtoLowSetLine();})
                         .addTemporalMarker(0.5, ()-> {encoded.openTopClaw();})
                         .addTemporalMarker(0.5,()-> {encoded.closeClaw();})
-                        .lineToConstantHeading(new Vector2d(42, 35.5)) // back up from bd
+                        .lineToConstantHeading(new Vector2d(43, 42)) // back up
                         .splineToConstantHeading(new Vector2d(60,9), Math.toRadians(0)) // spline into park (RIGHT)
                         //.splineToConstantHeading(new Vector2d(60,58.5), Math.toRadians(0)) // spline into park (LEFT)
-
 
                         .build();
 
@@ -60,40 +54,38 @@ public class rrBlueFrontTRUSSyellowOnly extends OpMode {
 
             case NONE:
             case MIDDLE:
-                drive.trajectorySequenceBuilder(new Pose2d(-35, 70, Math.toRadians(90)))
-                        .lineToConstantHeading(new Vector2d(-35, 34)) // to spikemark
+                drive.trajectorySequenceBuilder(new Pose2d(12, 74, Math.toRadians(90)))
+                        .lineToConstantHeading(new Vector2d(12,30)) // to spikemark
                         .addTemporalMarker(0,()-> {encoded.armtoGround();})
                         .addTemporalMarker(0.5,()->{encoded.openBottomClaw();})
-                        .lineToConstantHeading(new Vector2d(-42, 58.5)) // orientate + line up for truss
-                        .lineToConstantHeading(new Vector2d(42, 58.5)) // fly under truss
-                        .splineToLinearHeading(new Pose2d(50, 35.5), Math.toRadians(0)) // to bd
+                        .lineToConstantHeading(new Vector2d(12, 35)) // back up
+                        .splineToLinearHeading(new Pose2d(50,36), Math.toRadians(0)) // to bd
                         .addTemporalMarker(0,()-> {encoded.armtoLowSetLine();})
                         .addTemporalMarker(0.5, ()-> {encoded.openTopClaw();})
                         .addTemporalMarker(0.5,()-> {encoded.closeClaw();})
-                        .lineToConstantHeading(new Vector2d(42, 35.5)) // back up from bd
+                        .lineToConstantHeading(new Vector2d(43, 36)) // back up
                         .splineToConstantHeading(new Vector2d(60,9), Math.toRadians(0)) // spline into park (RIGHT)
                         //.splineToConstantHeading(new Vector2d(60,58.5), Math.toRadians(0)) // spline into park (LEFT)
 
                         .build();
-
-
                 break;
 
             case RIGHT:
-                drive.trajectorySequenceBuilder(new Pose2d(-35, 70, Math.toRadians(90)))
-                        .lineToConstantHeading(new Vector2d(-46, 38)) // to right spikemark
+                drive.trajectorySequenceBuilder(new Pose2d(12, 74, Math.toRadians(90)))
+                        .lineToConstantHeading(new Vector2d(23,31)) // safely move
+                        .lineToLinearHeading(new Pose2d(19,31, Math.toRadians(0))) // orientate
+                        .lineToConstantHeading(new Vector2d(9, 30)) // right spikemark
                         .addTemporalMarker(0,()-> {encoded.armtoGround();})
                         .addTemporalMarker(0.5,()->{encoded.openBottomClaw();})
-                        .lineToConstantHeading(new Vector2d(-46, 49)) // backup
-                        .lineToLinearHeading(new Pose2d(-46, 58.5, Math.toRadians(0))) // orientate + line up for truss
-                        .lineToConstantHeading(new Vector2d(42, 58.5)) // fly under truss
-                        .splineToLinearHeading(new Pose2d(50, 28.5), Math.toRadians(0)) // to bd
+                        .lineToConstantHeading(new Vector2d(12, 30)) // back up
+                        .splineToLinearHeading(new Pose2d(50,29), Math.toRadians(0)) // to bd
                         .addTemporalMarker(0,()-> {encoded.armtoLowSetLine();})
                         .addTemporalMarker(0.5, ()-> {encoded.openTopClaw();})
                         .addTemporalMarker(0.5,()-> {encoded.closeClaw();})
-                        .lineToConstantHeading(new Vector2d(42, 28.5)) // back up from bd
+                        .lineToConstantHeading(new Vector2d(43, 29)) // back up
                         .splineToConstantHeading(new Vector2d(60,9), Math.toRadians(0)) // spline into park (RIGHT)
                         //.splineToConstantHeading(new Vector2d(60,58.5), Math.toRadians(0)) // spline into park (LEFT)
+
                         .build();
                 break;
         }
@@ -103,5 +95,5 @@ public class rrBlueFrontTRUSSyellowOnly extends OpMode {
     public void loop() {
 
     }
-}
 
+}
