+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/rrautos/rrBlueFrontGATEyellowOnly.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.autos;
+package org.firstinspires.ftc.teamcode.rrautos;
 
 import com.acmerobotics.roadrunner.geometry.Pose2d;
 import com.acmerobotics.roadrunner.geometry.Vector2d;
@@ -11,8 +11,8 @@ import org.firstinspires.ftc.teamcode.main.Encoded;
 import org.firstinspires.ftc.teamcode.vision.FirstVisionProcessor;
 import org.firstinspires.ftc.vision.VisionPortal;
 
-@Autonomous(name = "rrBlueBackYellowOnly")
-public class rrBlueBackYellowOnly extends OpMode {
+@Autonomous(name = "RR BlueFrontGATEyellowOnly")
+public class rrBlueFrontGATEyellowOnly extends OpMode {
     private FirstVisionProcessor visionProcessor;
     private VisionPortal visionPortal;
     private Encoded encoded;
@@ -23,6 +23,7 @@ public class rrBlueBackYellowOnly extends OpMode {
         visionProcessor = new FirstVisionProcessor();
         visionPortal = VisionPortal.easyCreateWithDefaults(hardwareMap.get(WebcamName.class, "Webcam 1"), visionProcessor);
     }
+
     @Override
     public void init_loop() {
         telemetry.addData("Identified", visionProcessor.getSelection());
@@ -35,16 +36,20 @@ public class rrBlueBackYellowOnly extends OpMode {
         telemetry.addData("Identified", visionProcessor.getSelection());
         switch (visionProcessor.getSelection()) {
             case LEFT:
-                drive.trajectorySequenceBuilder(new Pose2d(12, 74, Math.toRadians(90)))
-                        .lineToConstantHeading(new Vector2d(23,35)) // to spikemark
+                drive.trajectorySequenceBuilder(new Pose2d(-35, 70, Math.toRadians(90)))
+                        .lineToConstantHeading(new Vector2d(-40, 50)) // positioning
+                        .lineToLinearHeading(new Pose2d(-37, 29, Math.toRadians(180))) // orientation
+                        .lineToConstantHeading(new Vector2d(-35, 29)) //slow push to spikemark
                         .addTemporalMarker(0,()-> {encoded.armtoGround();})
                         .addTemporalMarker(0.5,()->{encoded.openBottomClaw();})
-                        .lineToConstantHeading(new Vector2d(23, 40)) // back up
-                        .splineToLinearHeading(new Pose2d(50,42), Math.toRadians(0)) // to bd
+                        .lineToConstantHeading(new Vector2d(-38, 29)) // safe backup
+                        .lineToConstantHeading(new Vector2d(-49,11)) // line up for gate
+                        .lineToConstantHeading(new Vector2d(42, 11)) // fly under gate
+                        .splineToLinearHeading(new Pose2d(50,41.5), Math.toRadians(0)) // to bd
                         .addTemporalMarker(0,()-> {encoded.armtoLowSetLine();})
                         .addTemporalMarker(0.5, ()-> {encoded.openTopClaw();})
                         .addTemporalMarker(0.5,()-> {encoded.closeClaw();})
-                        .lineToConstantHeading(new Vector2d(43, 42)) // back up
+                        .lineToConstantHeading(new Vector2d(43, 41.5)) // back up
                         .splineToConstantHeading(new Vector2d(60,9), Math.toRadians(0)) // spline into park (RIGHT)
                         //.splineToConstantHeading(new Vector2d(60,58.5), Math.toRadians(0)) // spline into park (LEFT)
 
@@ -54,35 +59,39 @@ public class rrBlueBackYellowOnly extends OpMode {
 
             case NONE:
             case MIDDLE:
-                drive.trajectorySequenceBuilder(new Pose2d(12, 74, Math.toRadians(90)))
-                        .lineToConstantHeading(new Vector2d(12,30)) // to spikemark
+                drive.trajectorySequenceBuilder(new Pose2d(-35, 70, Math.toRadians(90)))
+                        .lineToConstantHeading(new Vector2d(-33.5,32)) // to middle spikemark
                         .addTemporalMarker(0,()-> {encoded.armtoGround();})
-                        .addTemporalMarker(0.5,()->{encoded.openBottomClaw();})
-                        .lineToConstantHeading(new Vector2d(12, 35)) // back up
-                        .splineToLinearHeading(new Pose2d(50,36), Math.toRadians(0)) // to bd
+                        .addTemporalMarker(0.5,()->{encoded.openBottomClaw();})                            .lineToConstantHeading(new Vector2d(-35, 41)) // backup
+                        .lineToLinearHeading(new Pose2d(-49,41, Math.toRadians(0))) // out the way of spikemark
+                        .lineToConstantHeading(new Vector2d(-49,11)) // line up for gate
+                        .lineToConstantHeading(new Vector2d(42, 11)) // fly under gate
+                        .splineToLinearHeading(new Pose2d(50,35), Math.toRadians(0)) // to bd
                         .addTemporalMarker(0,()-> {encoded.armtoLowSetLine();})
                         .addTemporalMarker(0.5, ()-> {encoded.openTopClaw();})
                         .addTemporalMarker(0.5,()-> {encoded.closeClaw();})
-                        .lineToConstantHeading(new Vector2d(43, 36)) // back up
+                        .lineToConstantHeading(new Vector2d(43, 35)) // back up
                         .splineToConstantHeading(new Vector2d(60,9), Math.toRadians(0)) // spline into park (RIGHT)
                         //.splineToConstantHeading(new Vector2d(60,58.5), Math.toRadians(0)) // spline into park (LEFT)
 
                         .build();
+
                 break;
 
             case RIGHT:
-                drive.trajectorySequenceBuilder(new Pose2d(12, 74, Math.toRadians(90)))
-                        .lineToConstantHeading(new Vector2d(23,31)) // safely move
-                        .lineToLinearHeading(new Pose2d(19,31, Math.toRadians(0))) // orientate
-                        .lineToConstantHeading(new Vector2d(9, 30)) // right spikemark
+                drive.trajectorySequenceBuilder(new Pose2d(-35, 70, Math.toRadians(90)))
+                        .lineToConstantHeading(new Vector2d(-46,38)) // to right spikemark
                         .addTemporalMarker(0,()-> {encoded.armtoGround();})
                         .addTemporalMarker(0.5,()->{encoded.openBottomClaw();})
-                        .lineToConstantHeading(new Vector2d(12, 30)) // back up
-                        .splineToLinearHeading(new Pose2d(50,29), Math.toRadians(0)) // to bd
+                        .lineToConstantHeading(new Vector2d(-46, 49)) // backup
+                        .lineToLinearHeading(new Pose2d(-35,49, Math.toRadians(0))) // out the way of spikemark
+                        .lineToConstantHeading(new Vector2d(-35,11))
+                        .lineToConstantHeading(new Vector2d(42,11))
+                        .splineToLinearHeading(new Pose2d(50,28.5), Math.toRadians(0)) // to bd
                         .addTemporalMarker(0,()-> {encoded.armtoLowSetLine();})
                         .addTemporalMarker(0.5, ()-> {encoded.openTopClaw();})
                         .addTemporalMarker(0.5,()-> {encoded.closeClaw();})
-                        .lineToConstantHeading(new Vector2d(43, 29)) // back up
+                        .lineToConstantHeading(new Vector2d(43, 28.5)) // back up
                         .splineToConstantHeading(new Vector2d(60,9), Math.toRadians(0)) // spline into park (RIGHT)
                         //.splineToConstantHeading(new Vector2d(60,58.5), Math.toRadians(0)) // spline into park (LEFT)
 
@@ -95,5 +104,5 @@ public class rrBlueBackYellowOnly extends OpMode {
     public void loop() {
 
     }
-
 }
+
