+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/rrautos/rrbluejon.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.autos;
+package org.firstinspires.ftc.teamcode.rrautos;
 
 import com.acmerobotics.roadrunner.geometry.Pose2d;
 import com.acmerobotics.roadrunner.geometry.Vector2d;
@@ -13,7 +13,7 @@ import org.firstinspires.ftc.teamcode.trajectorysequence.TrajectorySequence;
 import org.firstinspires.ftc.teamcode.vision.FirstVisionProcessor;
 import org.firstinspires.ftc.vision.VisionPortal;
 
-@Autonomous(name = "rrBlueBACKJon")
+@Autonomous(name = "RR BlueBACKJon")
 public class rrbluejon extends LinearOpMode{
     private Encoded encoded;
     private FirstVisionProcessor visionProcessor;
